EN6-173 Extract RichTextEditor into own class component
